requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
pymongo==4.6.0
dnspython==2.4.2
mongoengine==0.29.1
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
